+++ lib/model/common/src/pool/spec.ts
@@ -1,5 +1,6 @@
 import { Branded } from '../branding';
 import { PColumn, PColumnSpec } from '../drivers';
+import { ResultPoolEntry } from './entry';
 
 /** Any object exported into the result pool by the block always have spec attached to it */
 export interface PObjectSpec {
@@ -37,24 +38,26 @@ export function isPColumn<T>(obj: PObject<T>): obj is PColumn<T> {
   return isPColumnSpec(obj.spec);
 }
 
+export function isPColumnSpecResult(
+  r: ResultPoolEntry<PObjectSpec>
+): r is ResultPoolEntry<PObjectSpec> {
+  return isPColumnSpec(r.obj);
+}
+
+export function isPColumnResult<T>(
+  r: ResultPoolEntry<PObject<T>>
+): r is ResultPoolEntry<PColumn<T>> {
+  return isPColumnSpec(r.obj.spec);
+}
+
 export function ensurePColumn<T>(obj: PObject<T>): PColumn<T> {
-  if (!isPColumn(obj))
-    throw new Error(`not a PColumn (kind = ${obj.spec.kind})`);
+  if (!isPColumn(obj)) throw new Error(`not a PColumn (kind = ${obj.spec.kind})`);
   return obj;
 }
 
-export function mapPObjectData<D1, D2>(
-  pObj: PColumn<D1>,
-  cb: (d: D1) => D2
-): PColumn<D2>;
-export function mapPObjectData<D1, D2>(
-  pObj: PObject<D1>,
-  cb: (d: D1) => D2
-): PObject<D2>;
-export function mapPObjectData<D1, D2>(
-  pObj: PObject<D1>,
-  cb: (d: D1) => D2
-): PObject<D2> {
+export function mapPObjectData<D1, D2>(pObj: PColumn<D1>, cb: (d: D1) => D2): PColumn<D2>;
+export function mapPObjectData<D1, D2>(pObj: PObject<D1>, cb: (d: D1) => D2): PObject<D2>;
+export function mapPObjectData<D1, D2>(pObj: PObject<D1>, cb: (d: D1) => D2): PObject<D2> {
   return {
     ...pObj,
     data: cb(pObj.data)