+++ sdk/ui-vue/src/components/PlAgDataTable/sources/table-source-heterogeneous.ts
@@ -1,4 +1,4 @@
-import type { ColDef, IDatasource, RowModelType } from '@ag-grid-community/core';
+import type { ColDef, IServerSideDatasource, RowModelType } from '@ag-grid-community/core';
 import type { AxisSpec, PColumnSpec, PTableColumnSpec, PTableShape, PTableVector, PVectorDataString } from '@platforma-sdk/model';
 import { getAxisId } from '@platforma-sdk/model';
 import canonicalize from 'canonicalize';
@@ -78,7 +78,7 @@ export function updatePFrameGridOptionsHeterogeneousAxes(
   indices: number[],
 ): {
   columnDefs: ColDef[];
-  datasource?: IDatasource;
+  serverSideDatasource?: IServerSideDatasource;
   rowModelType: RowModelType;
   rowData?: unknown[];
 } {