+++ sdk/ui-vue/src/components/PlAgDataTable/sources/table-source.ts
@@ -1,4 +1,4 @@
-import type { ColDef, GridApi, IDatasource, IGetRowsParams, RowModelType } from '@ag-grid-community/core';
+import type { ColDef, IServerSideDatasource, IServerSideGetRowsParams, RowModelType } from '@ag-grid-community/core';
 import type { AxisId, JoinEntry, PColumnIdAndSpec, PFrameHandle, PObjectId } from '@platforma-sdk/model';
 import {
   type PColumnSpec,
@@ -16,7 +16,7 @@ import {
 } from '@platforma-sdk/model';
 import canonicalize from 'canonicalize';
 import * as lodash from 'lodash';
-import { type PlDataTableSheet } from '../types';
+import type { PlDataTableSheet } from '../types';
 import { getHeterogeneousColumns, updatePFrameGridOptionsHeterogeneousAxes } from './table-source-heterogeneous';
 
 /**
@@ -273,13 +273,12 @@ const isLabelColumn = (col: PTableColumnSpec) => col.type === 'column' && col.sp
  * Calculate GridOptions for p-table data source type
  */
 export async function updatePFrameGridOptions(
-  gridApi: GridApi,
   pfDriver: PFrameDriver,
   pt: PTableHandle,
   sheets: PlDataTableSheet[],
 ): Promise<{
   columnDefs: ColDef[];
-  datasource?: IDatasource;
+  serverSideDatasource?: IServerSideDatasource;
   rowModelType: RowModelType;
   rowData?: unknown[];
 }> {
@@ -329,49 +328,52 @@ export async function updatePFrameGridOptions(
     return updatePFrameGridOptionsHeterogeneousAxes(hColumns, ptShape, columnDefs, await pfDriver.getData(pt, indices), fields, indices);
   }
 
-  let lastParams: IGetRowsParams | undefined = undefined;
-  const datasource = {
-    rowCount,
-    getRows: async (params) => {
-      gridApi.setGridOption('loading', true);
+  let lastParams: IServerSideGetRowsParams | undefined = undefined;
+  const serverSideDatasource = {
+    getRows: async (params: IServerSideGetRowsParams) => {
       try {
         if (rowCount == 0) {
-          params.successCallback([], rowCount);
-          gridApi.setGridOption('loading', false);
-          gridApi.showNoRowsOverlay();
+          params.success({ rowData: [], rowCount });
+          params.api.showNoRowsOverlay();
+          params.api.setGridOption('loading', false);
           return;
         }
 
         // this is to avoid double flickering when underlying table is changed
-        if (lastParams && !lodash.isEqual(lastParams.sortModel, params.sortModel)) {
+        if (lastParams && !lodash.isEqual(lastParams.request.sortModel, params.request.sortModel)) {
           lastParams = undefined;
-          params.failCallback();
+          params.fail();
+          params.api.setGridOption('loading', true);
           return;
         }
         lastParams = params;
 
-        const length = params.endRow - params.startRow;
+        let length = 0;
         let rowData: unknown[] = [];
-        if (length > 0) {
-          const data = await pfDriver.getData(pt, indices, {
-            offset: params.startRow,
-            length,
-          });
-          rowData = columns2rows(fields, data);
+        if (rowCount > 0 && params.request.startRow !== undefined && params.request.endRow !== undefined) {
+          length = Math.min(rowCount, params.request.endRow) - params.request.startRow;
+          if (length > 0) {
+            const data = await pfDriver.getData(pt, indices, {
+              offset: params.request.startRow,
+              length,
+            });
+            rowData = columns2rows(fields, data);
+          }
         }
 
-        params.successCallback(rowData, ptShape.rows);
-        gridApi.setGridOption('loading', false);
-        gridApi.autoSizeAllColumns();
+        params.success({ rowData, rowCount });
+        params.api.autoSizeAllColumns();
+        params.api.setGridOption('loading', false);
       } catch (error: unknown) {
-        params.failCallback();
+        params.fail();
+        params.api.setGridOption('loading', true);
       }
     },
-  } satisfies IDatasource;
+  } satisfies IServerSideDatasource;
 
   return {
-    rowModelType: 'infinite',
+    rowModelType: 'serverSide',
     columnDefs,
-    datasource,
+    serverSideDatasource,
   };
 }