+++ etc/block-components/enter-numbers/model/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @milaboratories/milaboratories.test-enter-numbers.model
 
+## 1.0.14
+
+### Patch Changes
+
+- @platforma-sdk/model@1.7.16
+
 ## 1.0.13
 
 ### Patch Changes