+++ sdk/workflow-tengo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platforma-sdk/workflow-tengo
 
+## 1.6.3
+
+### Patch Changes
+
+- ddb92ea: exec: rename cache -> cacheMillis, cacheMinutes, cacheSeconds etc; deprecate .cache()
+
+## 1.6.2
+
+### Patch Changes
+
+- ef02b23: PFrames conversion spec update
+
 ## 1.6.1
 
 ### Patch Changes