+++ tests/config-local-ml-integration/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @milaboratories/config-local-ml-integration
 
+## 1.1.31
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.14.4
+
+## 1.1.30
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.14.3
+
+## 1.1.29
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.14.2
+
+## 1.1.28
+
+### Patch Changes
+
+- Updated dependencies [ce6d188]
+  - @milaboratories/pl-middle-layer@1.14.1
+
 ## 1.1.27
 
 ### Patch Changes