+++ tools/package-builder/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @platforma-sdk/package-builder
 
+## 2.10.3
+
+### Patch Changes
+
+- 6299770: Use HeadObject to check we have access to bucket
+
+## 2.10.2
+
+### Patch Changes
+
+- fb9ed15: Teach tengo builder to find assets and detect asset imports
+
+## 2.10.1
+
+### Patch Changes
+
+- f2a5f70: Correct way to handle S3 init exceptions. Print all unhandled error stacktraces to command output
+
 ## 2.10.0
 
 ### Minor Changes